Autmoate admin task
If the provider has verified all the information provider by the client
we want to mark the task as passed. If the provider hasn't verified all
the information from the client we want to mark the task as failed.
Tasks are associated with a `DfeSignIn::User` so we need to create one
of these for the provider. NOTE that the find_or_create includes the
organisation check, we may have providers with multiple organisations,
in such cases we'll create a duplicate record for them. The
DfeSignIn::User only has a single organisation so if we want to record
the organisation the user was logged in as we can't really avoid
duplicate users here.