Add button to resend provider verification email
The provider can only start the verification process via the link in the
email. If the provider has lost this email we need a mechanisim for the
ops team to resend the email.
As this admin logic is specific only to the FE policy we've opted to
move the controller into it's own namespace