Record who sent the email to the provider
The figma designs require us to display the admin who resent the
provider verification email. To capture this information we use the
existing note mechanism on claims. This is a bit fragile as we assume
the only "provider_verification" note is for emails being sent. It seems
like the label on notes maps to a task name, so we went with that for
the label. If in future we can take other provider verification notes we
may want to consider renaming this email specific label.