Move rendering answers to policy specific partial
We'll want to show different claims answers sections for FE claims along
with potentially different width tables.
We're moving the rendering of the body of the claim show page to policy
specific partials so we don't have a ton of conditionals. Already have
been able to remove an ECP / LUP check.