Show details on FE claims
Updates the admin claim show page to show the details of FE claims.
The designs for this page call for a full width govuk-container. Also
some of the values in the `dl`s show multiple answers, for these we
return an array of answers which will be rendered with <br><br> between
each answer.