Set heading width and show decision status
Not sure how I feel about this. I was considering extracting a policy
specific partial for the header but seemed like it was repeating a lot
of information just to change the width of the header container. Not a
fan of the approach taken here but thought it better than extrating a
new partial. Maybe something like `@claim.policy.has_a_long_name?` would
e better to branch on ¯\_(ツ)_/¯