Match across all policies
Requirement is to check for matching claims across all policies, except
for IRP, which should only flag matches with other IRP policies.

At the minute the only attribute we use on eligibility when checking for
matches is TRN, which is present in all the eligibilities we compare, if
however we start checking other attributes the db will throw an error as
the column may not be present on all eligibilities, hence having to
update the "matching_claims - blank trn" test.