Set the award amount on the claim
Part of a series of commits migrating the award amount on to the claim.
This commit adds the award amount column to the claim and updates the
code to set the award amount. We're still setting the award amount on
the eligibility, in a future commit the award amount will be removed
from the eligibility.