Refactor query
Refactors the CTE to build a generic eligibilities table we can join to.
Now when we add a new policy the payroll query should automatically pick
it up. Ideally we'd rename the "student_loan_repayment_amount" column in
the DB but that's quite involved, so for the time being we define an
'award_amount_column' column on the policy which returns the name of the
DB column we need to check to get the award amount.