Fix duplciate claims in list
The awaiting provider verification scope joins to notes, claims can have
more than one note, if they do duplicate claims are returned from the
query. To resolve this we call distinct on the claim id. Claims have a
couple of JSON columns so we can't use the `distinct` method directly,
instead we have to explicitly use distinct with the claim.id. We also
need to wrap the whole query in a where clause so when we chain on a
`count` rails generates the correct SQL rather than something like
`SELECT COUNT(count_column) FROM (SELECT DISTINCT ON (claims.id), claims.id)`
Additionally we have to make sure to include `submitted_at` in the
distinct clause as some of the scopes used to build the query include
order by submitted_at clauses.