Move setting slug to after action
Adjusts the completed slugs method to only update the list of visited
slugs if the form is valid.

There's been some rollbar errors where the user seems to have skipped
ahead in the journey but hasn't be redirected back by the
check_page_is_in_sequence callback.

All the tests pass but this seems like it could be risky. I'm not 100%
sure that no code is subtly relying on the visited slugs being updated
on failed form verifications.

We may want to consider a different approach to wizard state management,
maybe passing the session answers to a state machine to determine which
slugs are accessible.