Set verified at on the claim
We need to be able to access the timestamp a claim was verified in big
query. The provider verification is flagged in the analytics block list
so isn't sent over to big query (it's a json structure and some of it's
fields are pii). To ensure big query has access to the verification
timestamp we have added a new column on the claim `Claim#verified_at`
which will be sent over to big query. A follow up PR will back fill this
column on already verified claims.