Allow policies to overwrite admin task views
Allow a policy to define it's own view for a specific task.
If a policy defines a template `admin/tasks/<policy name>/tasks/<task
name>` we'll use that when rendering the task view. If there isn't a
policy specific view we'll fall back to the task view in
`admin/tasks/<taske name>`.