Capture practitioner names
When the provider enters the practitioner name it's stored in the
`Claim#first_name` and `Claim#surname` fields. When the practitioner
completes their portion of the journey the details they entered on the
personal details step overwrites the values entered by the provider in
the `Claim#first_name` and `Claim#surname` fields.
We have a requirement to show and compare both the provider
entered claimant name and practitioner entered claimant name in the
admin area, this commit stores the provider entered name information in
two new fields allowing us to do this.