Move payroll run to background job
Creating the November payroll run timed out (October's too).
This commit moves creating the payment records into a background job.

We need to pass the claim ids and topup ids to the background job, this
is because once the admin has reviewed the claims that will be part of
this payroll run (shown in the `new` action) we can't include any claims
created in the interim. We may want to consider someother method of
indicating what claims should be part of this payroll run (create
pending payments? flag claims / topups as reviewed for payroll?) so we
don't have to pass such a large amount of information to the background
job.