Replace call to rake task
Calling the rake task on production throws an error as the rake tasks
aren't loaded. To resolve this, rather than invoking the rake task we'll
call the code the rake task calls.
The test is probably a bit OTT but we want to make sure updates to the
gem don't break the interface we're now using.

See
https://github.com/DFE-Digital/dfe-analytics/blob/212e0a7e7d1e27ba434080387b3f7b020f082b1a/lib/dfe/analytics/tasks/import_entities.rake#L18
for implementation of the `import_entity` task.