Send emails when claim approved
When an EY claim is approved we want to send an email to the
practitioner and the provider.
Mailers don't let us send two email from one action, so the approve
method enqueues jobs to send each of the emails.
I think it would probably be better for policies to be told a claim has
been approved and handle policy specific actions like sending emails in
the polciy, so rather than:
`@claimpolicy.mailer.approved(@claim).deliver_later`
we'd do
`@claim.policy.claim_approved(@claim)`
and let the policy determine the action to take (sending emails).
However this is a bit different to how we do things now, so for the time
being we'll just call the mailer, even it it's a bit odd and sends two
emails.