Move selectable_subject_symbols into the journey.
There was a lot of duplicated business rules in this method.
We were explicity checking for `nqt_in_academic_year_after_itt`, and if
it was absent we were falling back to only the lup subjects. The ECP
policy is ineligible for non `nqt_in_academic_year_after_itt` claims so
it will be filtered out when we check for
`potentially_still_eligible_policies`.
We returned an empty array if the `current_academic_year` was outside
the lup policy dates, however the LUP#subject_symbols method already
returns an empty array of available subjects for years outside the
policy.

There's some weirdness in the
`AdditionalPaymentsForTeaching.selectable_subject_symbols` method. If
we're dealing with a trainee teacher, we return a hard coded list of
subjects, however trainee teachers won't be eligible as the
`EligibilityCheckable#common_ineligible_attributes?` checks if the
claimant is a trainee teacher, and is thus ineligible. It seems weird to
return a list of subjects if the claimant is ineligible. This was the
existing behaviour so I've kept it in place but this likely needs
revisiting.