Move selectable_itt_years_for_claim_year from lib
As part of the refactor to remove ECP we want to move the policy
specific functionality into the respective polices. This commit moves
the `selectable_itt_years_for_claim_year` out of the lib class and into
each policy, with the journey delegating this method to any
potentially_still_eligible policies.