Remove trainee_teacher from common_ineligible_attributes
ECP `specific_ineligible_attributes?` already checks if
`trainee_teacher?` is set. For LUP trainee teachers can still be
eligible later, so we want to make sure we don't return ineligible in
that case.
This previously hasn't surfaced as an issue due to how the code was
called (a short circut specific to trainee teachers in the page
sequence) but we want to move some of the code around which exposes this
bug.