Remove AP specifc logic from page sequence
Previously the page sequence contained specific logic for handling an
additional payments sub journey (traniee teacher), and rather than
modifying the slug sequence just overwrote what is returned.
Additionally there was a condition in the slug sequence that caused the
whole thing to be swapped out if we were dealing with a trainee teacher.
This commit removes the additional payments specific handling from the
page sequence, leaving it journey agnostic, and updates the additional
payments slug sequence to move the trainee teacher sub journey branching
to the top level.