Remove falsely passed task
The `ClaimStudentLoanDetailsUpdater` errored, so the `student_loan_plan`
was not set on these claims, however `ClaimStudentLoanDetailsUpdater`
swallows errors so `AutomatedChecks::ClaimVerifiers::StudentLoanPlan`
still ran. `AutomatedChecks::ClaimVerifiers::StudentLoanPlan` looks for
the presence of student loan data in the student_loans_data table but
does not check the claim has a plan when creating the passed task.
This data migration removes the tasks, allowing a future student loan
import to set the plan type on these claims (as they'll now be awaiting
the slc plan task).