Move error handling
Moves the error handling out of the `ClaimStudentLoanDetailsUpdater` and
into the job that calls it.
When performing an SLC upload, if the `ClaimStudentLoanDetailsUpdater`
errors, and swallows the error, then we can get in a scenario where the
claim doesn't have student loan details set on it but the claim does
have a passing task. As the job uses the presence of the task to
determine which claims need their details updating, subsequent SLC data
uploads don't set the claim's details.
This commit moves the error handling to wrap both the
`ClaimStudentLoanDetailsUpdater` and the `AutomatedCheck`, so we still
preserve the original intention of allowing the rest of the import to go
ahead if one of the claims is throwing an error for some reason.