Set the missing plan types
After the SLC upload error (see CAPT-1964 for discussion) we need to set
the missing SLC plans on the following claims.
As these claims are from October the `StudentLoansData` table
doesn't have entries for all of them, so we've just added the list of
claims and their plan types (which we determine offline) as an array in
this data migration. Also as there aren't `StudentLoansData` entries we
can't use the `StudentLoanPlan` claim verifier class to set the
task & note, so we create the task and note ourselves.