Fix hint text
Previously the hint text was showing all subjects to claimants, even if
they were not eligible for those subjects, instead only show subjects
the claimant is eligible for. If the claimant is not eligible for any
subjects they will have been kicked out of the journey prior to this
screen.