ECP is ineligible if policy is closed
This is additional functionality, so may want splitting out into a
separate PR, but it was requested as part of this ticket.
If the current acadmic year is after the ECP policy close date, then
selecting an ECP only school should show the claimant the ineligible
screen.
There's a slight bit of trickyness, the ECP::Eligibility model has an
enum for itt academic year, that only allows values for dates within
the ECP policy, or nil, adding a date outside this range throws an
error. The page sequences calls the claim submission form which
constructs an eligibility to check if the journey is completed, if the
user chooses an itt date that's outside the ECP policy range (possible
for future academic years) constructing the ecp eligibility throws an
error. To avoid this we're using the new `validate` option on enum, so
rather than throw an error when adding a date out side the range it now
generates a validation error. Some of the tests needed updating to add a
valid itt academic year as now ecp eligibility isn't valid if it's
missing the itt year.