Adds approved claims failing qualification tasks
One gotcha that we might need to deal with in this report is the
`dqt_teacher_status`. When the qualification claim verifier runs is uses
the dqt record to set some notes on the claim. If the claim has a
populated `dqt_teacher_status` field it uses that to build the dqt
teacher record object, if not the status is fetched from the api,
however this information from the api is not persisted to the claim.
When generating the report we don't want to hit the dqt api for
potentially many claims, so if we're missing this dqt status we don't
include it in the report. Checking the current academic year's claims
there don't seem to be any claims that would be returned from this
report that are missing their `dqt_teacher_status`. If missing
`dqt_teacher_status` in this report is causing issues for the ops team,
we could consider parsing the claim notes to get this information.