Adds duplicate claims report
Adds the query to identify duplicate approved claims.
As per the MatchingAttributeFinder claims can either be duplicates on
claim attributes or on eligibility attributes. Certain eligibilities can
only be compared with other eligibilities determined by the policy's
`SomePolicy.policies_claimable` method. Eligibilities define the
attributes they should be compared with other eligibilities by in their
`SomePolicy.eligibility_matching_attributes` method, however some
policies that are comparable don't implement the attributes used for
finding matches (eg `EarlyYearsPayments` is in `EarlyCareerPayments`
`policies_claimable` list, but it doesn't implement
`teacher_reference_number` which is what `EarlyCareerPayments` uses to
determine duplicates). We have to do some fairly gnarl joins and unions
to compare eligibilities!

When comparing claims we search all approved claims in the current
academic year for duplicates, ignoring the claim's policy's
`policies_claimable` list, this is how MatchingAttributeFinder works so
we've kept that behaviour. For each group in the
`CLAIM_ATTRIBUTE_GROUPS_TO_MATCH` list we AND the conditions together,
normalising strings, we then use UNION to handle the OR conditions, this
performs significantly better than complicating the join condition. When
adding new attributes to the duplicate detection list we _really_ need
to make sure we add the appropriate indexes.