Remove policies_claimable
This constant was introduced in e25c92f to handle policies not having a
trn (part of the moving trn from claim to eligibility work).
806c202d8d introduced some logic to only compare eligibility attributes
that both policies have in common and so the policies_claimable constant
is redundant.
We always want to show a matching claim alert in the admin area even if
the policy combination is invalid (eg can't claim for both irp & fe),
thus there is no need to limit the policies we look for matches against.