Add duplicate claims check
We're soon going to want to generate a report for the ops team of the
duplicate claims. This commit introduces a new claim verifier that
records the duplicate claims, if any.
Once existing duplicates on production have been back filled we can
replace other uses of `MatchingAttributeFinder#matching_claims` with
`claim.duplicates` and also join claims to their duplicates when
generating the duplicate claims report.