Block invalid policy combinations
The business rules state that certain policy combinations aren't
permitted, eg if a claimant has made an ECP claim they can't make a
second claim for FE in the same academic year. In such an invalid claim
scenario the requirement is to disable the approve button so admins
can't approve the claim.