Log session answers to rollbar
Some of the errors in rollbar are tricky to debug as we don't know what
answers the claimant provided. This commit adds a mechanism for flagging
attributes as PII and updates the ClaimsController to include the non
pii attributes in the Rollbar scope such that they are available in
Rollbar.