Remove Claim::FILTER_PARAMS
Claim filter params is no longer needed as it's been superseded by the
`pii` option on the SessionAttribute answers.
As the wizards are backed by the session answers class and don't
directly write attributes to the claim we've decided to keep the
sensitive attribute flagging as close to the attribute definition as
possible, this removed the need for developers to remember to update
this list on the claim when writing adding a new attribute that might
not be written to the claim (it may be written to an eligibility).