Create audit trail when SLC upload runs
When the SLC data upload runs we want to use the existing amendments
mechanism to record any changes to the claim. As amending a claim
requires a user to be associated with the amendment we need to pass who
triggered the SLC upload down to the `ClaimStudentLoanDetailsUpdater`.

We've had to slightly change the validation on
`StudentLoans::Eligibility` to check if the repayment amount has changed
as there's one test (`spec/features/admin/upload_slc_data_spec.rb`)
where:
* No SLC data is found for a claim
* The claim has a student_loan_repayment_amount of `0`
* The claim's plan is expected to change from `not_applicable` to `nil`
* The claim's `has_student_loan` is expected to change from `false` to
  `nil`
Without changing the validation on the eligibility we can't pass this
test as amending the claim errors due to the `0` student loan repayment
amount.