Add top up confirmation
Makes the topup form multi step with a confirmation screen.
Once an admin has entered a topup amount we need to display a page to
confirm the amount before we create the topup.
We handle tracking the state by passing around a step param and using
contextual validations. If all validations pass we consider the form
complete and create the records.
We use `dup` in the `complete?` method to avoid rendering errors for
steps other than the one we're currently on.