Don't cache admin pages
We already set the max-age=0, private, must-revalidate directives but we
need to add no-store so the pages don't show up in the BF cache.
ITHC picked up an issue where after logging out of the admin area
hitting the back button renders a logged in page, seems like it's BF
cache doing this.