Rate limit file uploads
Feedback from the ITHC.
We don't have a cache set so rate limiting will use the file cache,
storing the data in ./tmp on the containers. The rate limiting will be
per container, rather than app wide but given these uploads are
restricted to admins it wont be an issue. We don't do a lot of file
uploads so storing this info in the container file system shouldn't be
an issue.
If we start supporting public file uploads we'll want to investigate a
different cache store.