Adds service access code
Service access codes are a mechanism for allowing claimants to access
the service outside of the claim window.
These links will be created adhoc by developers in the console

```rb
Journeys::ServiceAccessCode.create!(
  journey: Journeys::FurtherEducationPayments
)
```

Some of the journeys are started via an email link
* FE provider
* EY provider authenticated
* EY practition

For these journeys we can either
* open the journey as needed (should be ok as they're not generally
  accessible)
* manually construct the links to include a service access code and tell
  the users to ignore links received by email
* update the code to include service access codes in any emails if the
  emails were triggered by a session containing a service access code

For now we've not implemented any specific code to handle these journeys
as we may not need it / can get by with work arounds.