Add name / dob fields to analytics for FE only
We want to send over name and date of birth over to DfE analytics for FE
claims only. These attributes are stored on the claim which is shared
between all policies. DfE analaytics gem only gives us coarse grain
control over which attributes to send to big query. This is a hack and
we're not stoked on having to patch internal methods like this but it
should do the trick until we think of a better solution.