Add current_academic_year to policy
In quite a few places in the app we're getting the current academic year
from the policy by going through the journey. This commit adds a method
to the base policy to return the current academic year to avoid the
chained method calls, and make it easier to overwrite this method if
needs be.
Ideally we'd be able to use `AcademicYear.current` however there are
some scenarios, such as in QA,  where being able to run a journey in a
different acadmic year is useful.