Don't allow verifying rejected claims
If the claim has been rejected we don't want the provider to be able to
verify the cliam.
The FormSubmittable module assumes there is a next slug or we want to
define a callback if there isn't one. In our use case we don't want to
define a callback and there is no next slug, so the simplest work around
is to just have two "expired-link" slugs.