Add Form::Option
As part of the LUP separate journey work I've found myself introducing a
Struct to represent form options in a couple of forms. This commit adds
an Option class to the base form and replaces usage of OpenStruct with
it. We use a class rather than a Struct as we have some optional key
words.
`value` and `label` would probably make more sense as attribute names
but the existing code, and the form builder docs, use `id` and `name` so
we've stuck with them.