Pass reference to a session to answers
Sometimes we need a reference to the session in the answers for things
like timestamps.
Annoyingly the answers method needs to be defined on each subclass.
Ideally we could move the `answers` attribute and this method to the
base class, and lose the `SessionAnswerType` subclasses, but I'm not sure
how we could do this and still support `SessionAnswer` subclasses for
each journey.