Move setting answers attribute into the base class
This might be too magical.

We want to be able to make a reference to the session availble in the
answers model. This allows accessing things like timestamps from the
answers and also allows us to define save / update on the answers.

We need to overwrite the answers method on each subclass, as the answers
attribute is only defined on the subclasses, so we have to do this in
the inherited hook with define method.