Write answers to journey session first
Introduces a shim where we the verify claim form writes the answers to
the journey session first then calls the claim submission form to
persist the answers / triggers post submission actions.
We'll soon be introducing another step in the journey and this change
makes it easier to support additional forms.