Prefill form attributes from verification
Verification is a bit unusual as we store the verification in a json
data structure (we may want to visit this and just store flat
attributes) so the base form doesn't set the attributes from the journey
session. Now the provider journey is multi step we need to make sure
there's an attribute on the session answers for the
`assertions_attributes` such that the base form can prefill the
verifications.