Adds claim identity verification
For FE claims where OL IDV fails we want the provider verification
journey to include an extra form to verify the claimant's identity.
This commit adds the new page.
We store the attributes from this form as "flat" attributes on the
eligibility, this works better with the assumptions the app makes about
eligibility data, the json verification attributes are a bit trickier to
work with. We may want to consider migrating the verification attributes
to individual column attributes on the eligibility.