Add additional dob validations
We need to give some more feedback on the date of birth form. This is
tricky as we receive the date parts as multi parameter attributes.
We'd considered extracting this logic into a date of birth validator but
it's dependent on the how the form handles multi part date params.