Add alternative identity verification task
If a claimant fails one login IDV we need to ask the provider to verify
the claimants identity. This commit adds the new task to the admin area
to track this alternative identity verification.

When a claim is considered to be a duplicate we don't automatically send
the verification email to the provider, it needs to be manually
triggered by an ops team member. The designs call for showing the resend
verification email on both the `alternative_identity_verification` and
`provider_verification` task views, both designs for this are the same,
so we've reused the parital from the `provider_verification` that
already handles this.