Move task status logic
Moves the task status logic to it's own module.
This helper is getting a bit larger and we want to add some more
branches to support alternative identity verification tasks.
Originally we pulled this out into a class but it's really just a pure
function so it's better suited as a module level method.