Don't create task unless IDV is a pass
In order to support an upcoming piece of work where we don't allow
approving a claim if the alternative IDV task is still pending we need
to change how we create this task.

`ClaimCheckingTasks#incomplete_task_names` assumes that a task hasn't
been completed if the Task record for a given task name doesn't exist.
In order to reuse this functionality we now only create the task in the
claim verifier if the provider details match the claimant details. If
there is no match then the ops team need to manually complete the task,
so we can use the absence of the task record to indicate an outstanding
task.

We've updated the `Tasks.status` method to branch on the task name
rather than adding more conditionals to the generic case. At some point
we should consider splitting out the generic case a bit.