Add atl idv rejection reason
We want a new rejection reason for claims that have failed alternative
identity verification. We only want this rejection reason to be
avalabile after 31st March.
The tricky bit is any rejection reasons we add to the notify template
need to be sent for all rejection emails. To handle this we only include
the rejection reason in the list of reasons admins see if the claim
requires alternative identity verification (which is feature flagged)
however the `Decision#rejected_reasons_hash` which is what is sent to
notify unconditionally includes this reason.

The `FurtherEducationPayments` policy class now has quite a few methods
that take a `Claim` as their only argument, while this works fine and is
pretty maintainable, it hints that we're missing some object.