Add form for test users to set return codes
In environments where we're bypassing one login we now present a form to
testers where they can enter details that would be returned from one
login, along with setting the desired one login return codes.
Now, rather than failing IDV in test based on the `alternative_idv`
feature flag we fail/pass IDV based on the presence of return codes
entered in the form.

Due to some quirks of how the form builder works, we've had to reuse the
`PersonalDetailsForm` to support this functionality. We subclass the
`PersonalDetailsForm` in the controller as `OneLoginTestUser` to make
bypassing it in tests more straightforward.

I think a slightly nicer approach would be to update the
`test_user_auth_hash` method to return a JWT with the content we want,
so we only have to branch on `OneLoginSignIn.bypass?` in one place. This
however would require moving quite a bit more code (injecting
`OneLogin::DidCache` as an are to the `CoreIdentityValidator`
constructor). Something to consider if we decide to refactor this
controller when we're a bit less pressed for time.