Fix missing nursery name in report
The default scope on EligibleEyProvider means we can't find the provider
for claims in the following scenario:

* EY claim submitted for provider A
* EY provider CSV is updated to no longer include provider A

In the admin area in reports we need to show who the provider was even
if the eligible providers list has been updated.

To resolve this we're unscoping the default scope when looking for the
provider and picking the latest one with a matching urn.

We probably want to revisit our use of default scope as we'll
undoubtedly be playing whack-a-mole with unscoping it.
Additionally we should move `eligible_ey_provider` to a scope on the
eligibility, this will make queries easier to run, and we should look
into why we're using `find_by`, there should always be an ey provider
for a submitted claim, so we should use `find_by!` or equivalent.