Move no_selectable_subjects?
This method doesn't need to be shared with LUP as there will always be a
selectable LUP subject.

When calculating the subject_symbols for LUP/TRI we return an empty
array if the itt year is earlier than the previous 5 years, however we
only show the previous 5 years as options for selecting the itt year,
and if the user chooses "none of the above", they're kicked out of the
journey before this screen by the `ineligible_cohory?` check.

`Policies::TargetedRetentionIncentivePayments.subject_symbols` should
just reduce to
`Policies::TargetedRetentionIncentivePayments.fixed_subject_symbols` but
we'll refactor towards that in a separate commit.