Get provider from eligibility
The default scope on EligibleEyProvider causes us to lose access to EY
providers after a file upload that replaces them. This is a quick fix to
unblock the admin area (we unscope eligibile ey provider on the
eligibility). We should probably remove the default scope on
EligibleEyProvider and add a new scope `active` or something to return
the EY providers from the latest file upload and use that scope in the
journey where we want to restrict the list of EY providers to those in
the latest file upload.