Update backlink spec to work with TRI only
We're removing the additional payments journey so this spec needs
updating to use the TRI only journey.

TRI only journey uses the navigator rather than the page sequence so it
doesn't handle the dead end slugs. We can't update the navigator to
return `nil` for the previous slug if the claimant is ineligible as some
FE pages use `previous_slug` on the ineligible page. For the time being
we'll check the page seqeuence's dead end slugs when setting the
backlink path but this is a bit of a hack and needs revisiting when we
switch all journeys to use the navigator.