Handle switching journeys
This is a bit of a hack. If the current slug passed to the navigator is
`nil` it assumes we're starting a new journey and returns the first
slug, however in the case of switching journey we want the user to go to
the next form the need to complete. The navigator doesn't care about the
slug passed to it if there are outstanding forms, so we just give it a
place holder.
We likely want to think up a better solution than this hack!
I think we should probably remove the "session in progress"
functionality as we don't needed it, it was put into the app at a time
where sessions weren't namespaced by journey and so you couldn't have
two in-flight claims at once. We now namespaces sessions by journey an
this "session in progress" functionality is redundant.