Migrate TRI reminders
We're removing the additional payments journey and replacing it with a
targeted retention incentive payments only journey. There are 3296
claimants who are eligibile for a TRI claim next year who set a reminder
after completing the additional payments journey. We want to ensure
these claimants still get a reminder when we open the new TRI only
journey.