Adds previous year check for IRP
A requirement for IRP is to flag to admins if a new claim matches a
claim in a previous year. We match on national insurance number which,
for IRP, is retained for 2 years. Once the 2 year retention period for
NINO is passed we'll be scrubbing this data and so we need a more
permanent association between claims, hence the new join model
`Claims::Match` which associates two claims together.