Validate matches are from previous year
The claim match class is pretty generic and the admin view code assumes
the match is always from the previous year. We're adding this validation
to protect the assumption in the view.
If this code is being changed to support other use cases then we
probably want to add some scopes for finding previous year matches and
updated `app/views/admin/tasks/first_year_payment.html.erb` to use these
scopes.